bcache: Fix an accounting integer overflow
We store the various statistics in long integers, but those are updated
on a timer from 32 bit atomic integers. If that 32 bit int wraps around
in the ~13 seconds between timer intervals, it'd end up adding a very
large negative number - since atomic_t is signed it'd get sign
into a signed long, then converted to unsigned long... not what we want.

Convert it to unsigned int first, so we get the full 32 bits and if it
wraps that we get 0 instead of ULONG_MAX.

+++ drivers/md/bcache/stats.c
@@ -169,7 +169,7 @@ static void scale_accounting(unsigned long data)
 	struct cache_accounting *acc = (struct cache_accounting *) data;
 
 #define move_stat(name) do {						\
-	unsigned long t = atomic_xchg(&acc->collector.name, 0);		\
+	unsigned t = atomic_xchg(&acc->collector.name, 0);		\
 	t <<= 16;							\
 	acc->five_minute.name += t;					\
 	acc->hour.name += t;						\
